+++ Ecommerce-api/src/main/java/gov/iti/jets/persistence/repository/customer/CustomerRepository.java
@@ -11,13 +11,11 @@ import jakarta.persistence.EntityManagerFactory;
 import jakarta.persistence.Persistence;
 import jakarta.persistence.TypedQuery;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 public class CustomerRepository {
-    EntityManagerFactory emf = Persistence.createEntityManagerFactory("api");
+    EntityManagerFactory emf = Persistence.createEntityManagerFactory( "api" );
     EntityManager em;
 
     public List<Customer> getAllCustomers() {
@@ -43,29 +41,35 @@ public class CustomerRepository {
         em = emf.createEntityManager();
         em.getTransaction().begin();
         Customer customer = em.find( Customer.class, id );
+        if(customer ==  null){
+            return new ArrayList<>();
+        }
         em.getTransaction().commit();
         em.close();
-        return customer.getOrders().stream().collect( Collectors.toList());
+        return customer.getOrders().stream().collect( Collectors.toList() );
     }
 
     public Order getOrderByIdOfCustomer( int id, int oid ) {
         em = emf.createEntityManager();
         em.getTransaction().begin();
         Customer customer = em.find( Customer.class, id );
-        for ( Order order: customer.getOrders() ) {
-            if ( order.getId() == oid ){
-                return order;
-            }
+        if(customer == null){
+            return null;
         }
         em.getTransaction().commit();
         em.close();
-        return null;
+        return customer.getOrders().stream().filter( order -> {
+            return order.getId() == oid;
+        } ).findFirst().get();
     }
 
-    public String  deleteAllCustomers() {
+    public String deleteAllCustomers() {
         em = emf.createEntityManager();
         em.getTransaction().begin();
         TypedQuery<Customer> customerQ = em.createQuery( "from Customer c", Customer.class );
+        if(customerQ.getResultList().isEmpty()){
+            return null;
+        }
         for ( Customer customer : customerQ.getResultList() ) {
             em.remove( customer );
         }
@@ -79,40 +83,49 @@ public class CustomerRepository {
         em = emf.createEntityManager();
         em.getTransaction().begin();
         Customer customer = em.find( Customer.class, id );
+        if(customer == null){
+            return null;
+        }
         em.remove( customer );
         em.getTransaction().commit();
         em.close();
-        return "Customer with id "+id+" was deleted successfully";
+        return "Customer with id " + id + " was deleted successfully";
     }
 
     public String addCustomer( CustomerPostRequest customerPostRequest ) {
         em = emf.createEntityManager();
         em.getTransaction().begin();
-        em.persist( new Customer(customerPostRequest.getName(),customerPostRequest.getEmail(),customerPostRequest.getPhone(),null) );
+        em.persist( new Customer( customerPostRequest.getName(), customerPostRequest.getEmail(), customerPostRequest.getPhone(), null ) );
         em.getTransaction().commit();
         em.close();
         return "Customer added successfully";
     }
 
-    public String addOrderToCustomer( CustomerOrderPostRequest customerOrderPostRequest , int id) {
+    public String addOrderToCustomer( CustomerOrderPostRequest customerOrderPostRequest, int id ) {
         em = emf.createEntityManager();
         em.getTransaction().begin();
         Set<Product> products = new HashSet<>();
+        Customer customer = em.find( Customer.class, id );
+        if(customer == null){
+            return null;
+        }
         for ( Integer productId :
                 customerOrderPostRequest.getProductsId() ) {
-            products.add( em.find( Product.class , productId ) );
+            products.add( em.find( Product.class, productId ) );
         }
-        Customer customer = em.find( Customer.class, id );
-        em.persist( new Order(customer,customerOrderPostRequest.getTotalPrice(),products) );
+        em.persist( new Order( customer, customerOrderPostRequest.getTotalPrice(), products ) );
         em.getTransaction().commit();
         em.close();
-        return "Order was added to customer with id "+id +" successfully";
+        return "Order was added to customer with id " + id + " successfully";
     }
 
     public Customer editCustomer( CustomerPostRequest customerPostRequest, int id ) {
         em = emf.createEntityManager();
         em.getTransaction().begin();
         Customer customer = em.find( Customer.class, id );
+        if(customer == null){
+            return null;
+        }
         customer.setName( customerPostRequest.getName() );
         customer.setEmail( customerPostRequest.getEmail() );
         customer.setPhone( customerPostRequest.getPhone() );
@@ -126,13 +139,21 @@ public class CustomerRepository {
         em = emf.createEntityManager();
         em.getTransaction().begin();
         Customer customer = em.find( Customer.class, id );
-        Order order = customer.getOrders().stream().filter( o -> {
+        if(customer == null){
+            return null;
+        }
+        Optional<Order> first = customer.getOrders().stream().filter( o -> {
             return o.getId() == oid;
-        } ).findFirst().get();
+        } ).findFirst();
+        if(!first.isPresent()){
+            return null;
+        }
+        Order order = first.get();
         order.setTotalPrice( customerOrderPostRequest.getTotalPrice() );
         Set<Product> products = new HashSet<>();
         customerOrderPostRequest.getProductsId().forEach( productId -> {
-            products.add( em.find( Product.class, productId ) );} );
+            products.add( em.find( Product.class, productId ) );
+        } );
         order.setProducts( products );
         order = em.merge( order );
         em.getTransaction().commit();
